+++ src/extension.ts
@@ -1,0 +1,24 @@
+'use strict';
+// The module 'vscode' contains the VS Code extensibility API
+// Import the module and reference it with the alias vscode in your code below
+import * as vscode from 'vscode';
+import { SystemVerilogDocumentSymbolProvider } from './providers/DocumentSymbolProvider';
+import { SystemVerilogWorkspaceSymbolProvider } from './providers/WorkspaceSymbolProvider';
+
+// this method is called when your extension is activated
+// your extension is activated the very first time the command is executed
+export function activate(context: vscode.ExtensionContext) {
+
+    context.subscriptions.push(vscode.languages.registerDocumentSymbolProvider(
+        {language: "systemverilog"}, new SystemVerilogDocumentSymbolProvider()
+    ));
+
+    context.subscriptions.push(vscode.languages.registerWorkspaceSymbolProvider(
+        new SystemVerilogWorkspaceSymbolProvider()
+    ));
+
+}
+
+// this method is called when your extension is deactivated
+export function deactivate() {
+}