+++ CHANGELOG.md
@@ -5,7 +5,13 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 - IntelliSense support
+- Back-end Language server for Systemverilog
+- Complete syntax highlighting
 
 ## [0.0.1] - 2018-01-15
-- Inital release
-- Syntax highligthing
+- Initial release
+- Initial Syntax highligthing (Not complete)
+
+## [0.0.2] - 2018-01-23
+- Go to symbol/module in workspace
+- Go to symbol in document